+++ static/static.go
@@ -9,16 +9,32 @@ import (
 	"golang.org/x/tools/go/ssa/ssautil"
 )
 
-// LoadPackagesConfigMode is the set of load mode values required to obtain the
-// information necessary to statically analyze Dogma applications.
-const LoadPackagesConfigMode = packages.NeedFiles |
+// PackagesLoadMode is the minimal [packages.LoadMode] required when loading
+// packages for analysis by [FromPackages].
+const PackagesLoadMode = packages.NeedFiles |
 	packages.NeedCompiledGoFiles |
 	packages.NeedImports |
 	packages.NeedTypes |
-	packages.NeedTypesSizes |
 	packages.NeedSyntax |
 	packages.NeedTypesInfo |
 	packages.NeedDeps
+
+// FromDir returns the configurations of the Dogma applications implemented
+// within the packages in the given directory and its children.
+func FromDir(dir string) []configkit.Application {
+	pkgs, err := packages.Load(
+		&packages.Config{
+			Mode: PackagesLoadMode,
+			Dir:  dir,
+		},
+		"./...",
+	)
+	if err != nil {
+		panic(err)
+	}
+
+	return FromPackages(pkgs)
+}
 
 // FromPackages returns the configurations of the Dogma applications implemented
 // within a set of packages.